docs: Add disclaimer about redundant module.
Honestly, you don't need to npm install this module.

Just understand how it works and use it in your apps, it's
just a couple of lines of glues code.

For example, we have an internal wrapper around `leveldown`
that adds Promise<{ err, data }> support to it instead of
using callbacks.

We also have a central place where we wrap the `aws-sdk` and
just do `try { r.promise() } catch (err) {}`.